+++ metldata/builtin_transformations/infer_references/main.py
@@ -1,0 +1,72 @@
+# Copyright 2021 - 2023 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
+# for the German Human Genome-Phenome Archive (GHGA)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+"""A transformation to infer references based on existing ones in the metadata model."""
+
+from metldata.builtin_transformations.infer_references.config import ReferenceMapConfig
+from metldata.builtin_transformations.infer_references.metadata_transform import (
+    transform_metadata,
+)
+from metldata.builtin_transformations.infer_references.model_transform import (
+    transform_metadata_model,
+)
+from metldata.model_utils.assumptions import check_basic_model_assumption
+from metldata.model_utils.essentials import MetadataModel
+from metldata.transform.base import Json, TransformationBase
+
+
+class ReferenceInferenceConfig(ReferenceMapConfig):
+    """Config parameters and their defaults."""
+
+
+class ReferenceInferenceTransformation(TransformationBase):
+    """A transformation to infer references based on existing ones in the metadata
+    model."""
+
+    def __init__(self, *, model: MetadataModel, config: ReferenceInferenceConfig):
+        """Initialize the transformation with transformation-specific config params and
+        the metadata model. The transformed model will be immediately available in the
+        `transformed_model` attribute (may be a property).
+
+        Raises:
+            MetadataModelAssumptionError:
+                if assumptions about the metadata model are not met.
+            MetadataModelTransformationError:
+                if the transformation of the metadata model fails.
+        """
+
+        self._original_model = model
+        self._inferred_references = config.inferred_references
+
+        check_basic_model_assumption(model=model)
+
+        self.transformed_model = transform_metadata_model(
+            model=self._original_model, references=self._inferred_references
+        )
+
+    def transform_metadata(self, metadata: Json) -> Json:
+        """Transforms metadata and returns it.
+
+        Raises:
+            MetadataTransformationError:
+                if the transformation of the metadata fails.
+        """
+
+        return transform_metadata(
+            metadata=metadata,
+            model=self.transformed_model,
+            references=self._inferred_references,
+        )
