+++ metldata/builtin_transformations/infer_references/path/path_elements.py
