+++ metldata/builtin_transformations/infer_references/reference.py
@@ -18,7 +18,7 @@
 
 from pydantic import BaseModel, Field, root_validator
 
-from metldata.reference.path import ReferencePath
+from metldata.builtin_transformations.infer_references.path.path import ReferencePath
 
 
 class ReferenceDetails(BaseModel):
